Debugging.  Generates neighborhood (lots of 1 and -1).

+++ Database/Primitives/User.java
@@ -38,7 +38,7 @@ public class User {
         sumRatings += song.getRating();
     }
     
-    public int getAvgRating(){
+    public double getAvgRating(){
         return sumRatings/ratings.getCount();
     }
